+++ routers/admin_api/v1/delivery.go
@@ -13,10 +13,10 @@ import (
 
 type DeliveryRuleParams struct {
 	ID            int         `json:"id,omitempty"`
-	First         float32     `json:"first" ` // 首件/首重
-	FirstFee      float32     `json:"first_fee" `
-	Additional    float32     `json:"additional" `     // 续件/续重
-	AdditionalFee float32     `json:"additional_fee" ` // 续件/续重
+	First         float64     `json:"first" ` // 首件/首重
+	FirstFee      float64     `json:"first_fee" `
+	Additional    float64     `json:"additional" `     // 续件/续重
+	AdditionalFee float64     `json:"additional_fee" ` // 续件/续重
 	Region        models.JSON `json:"region" `         // 可配送区域(省id集)
 	Destroy       bool        `json:"_destroy,omitempty"`
 	Position      int         `json:"position"`
